requirements.txt
customtkinter
pyautogui
pytesseract
openai==0.28.1
Pillow
python-dotenv
pyinstaller
pydantic
typing-extensions